+++ Sources/StreamVideo/StreamVideo.swift
@@ -14,6 +14,8 @@ public typealias UserTokenUpdater = (UserToken) -> Void
 /// Needs to be initalized with a valid api key, user and token (and token provider).
 public class StreamVideo: ObservableObject, @unchecked Sendable {
     
+    @Injected(\.callCache) private var callCache
+
     public class State: ObservableObject {
         @Published public internal(set) var connection: ConnectionStatus = .initialized
         @Published public internal(set) var user: User
@@ -143,14 +145,21 @@ public class StreamVideo: ObservableObject, @unchecked Sendable {
         // This is used from the `StreamCallAudioRecorder` to observe active
         // calls and activate/deactivate the AudioSession.
         StreamActiveCallProviderKey.currentValue = self
+
+        // Clear up the call cache to avoid stale call objects.
+        callCache.removeAll()
+
         (apiTransport as? URLSessionTransport)?.setTokenUpdater { [weak self] userToken in
             self?.token = userToken
         }
         if user.type != .anonymous {
-            let userAuth = UserAuth { [unowned self] in
-                self.token.rawValue
-            } connectionId: { [unowned self] in
-                await self.loadConnectionId()
+            let userAuth = UserAuth { [weak self] in
+                self?.token.rawValue ?? ""
+            } connectionId: { [weak self] in
+                guard let self else {
+                    throw ClientError.Unexpected()
+                }
+                return await self.loadConnectionId()
             }
             coordinatorClient.middlewares.append(userAuth)
         } else {
@@ -190,15 +199,17 @@ public class StreamVideo: ObservableObject, @unchecked Sendable {
         callType: String,
         callId: String
     ) -> Call {
-        let callController = makeCallController(callType: callType, callId: callId)
-        let call = Call(
-            callType: callType,
-            callId: callId,
-            coordinatorClient: coordinatorClient,
-            callController: callController
-        )
-        eventsMiddleware.add(subscriber: call)
-        return call
+        callCache.call(for: callCid(from: callId, callType: callType)) {
+            let callController = makeCallController(callType: callType, callId: callId)
+            let call = Call(
+                callType: callType,
+                callId: callId,
+                coordinatorClient: coordinatorClient,
+                callController: callController
+            )
+            eventsMiddleware.add(subscriber: call)
+            return call
+        }
     }
 
     /// Creates a controller used for querying and watching calls.